requirements.txt
Django>=1.8,<2.0
psycopg2
django-mptt
django-mptt-admin
django-multiselectfield
sorl-thumbnail
Pillow
django-nested-admin